fix(cas): package names aligned, log.warn removed in favor of println
