--- grails-test-examples/spring-security/cas/test1/grails-app/domain/com/test/User.groovy
@@ -1,79 +1,0 @@
-/*
- *  Licensed to the Apache Software Foundation (ASF) under one
- *  or more contributor license agreements.  See the NOTICE file
- *  distributed with this work for additional information
- *  regarding copyright ownership.  The ASF licenses this file
- *  to you under the Apache License, Version 2.0 (the
- *  "License"); you may not use this file except in compliance
- *  with the License.  You may obtain a copy of the License at
- *
- *    https://www.apache.org/licenses/LICENSE-2.0
- *
- *  Unless required by applicable law or agreed to in writing,
- *  software distributed under the License is distributed on an
- *  "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- *  KIND, either express or implied.  See the License for the
- *  specific language governing permissions and limitations
- *  under the License.
- */
-
-package com.test
-
-import groovy.transform.EqualsAndHashCode
-import groovy.transform.ToString
-
-@EqualsAndHashCode(includes='username')
-@ToString(includes='username', includeNames=true, includePackage=false)
-class User implements Serializable {
-
-	private static final long serialVersionUID = 1
-
-	transient springSecurityService
-
-	String username
-	String password
-	boolean enabled = true
-	boolean accountExpired
-	boolean accountLocked
-	boolean passwordExpired
-
-	User(String username, String password) {
-		this()
-		this.username = username
-		this.password = password
-	}
-
-	Set<Role> getAuthorities() {
-		UserRole.findAllByUser(this)*.role
-	}
-
-	def beforeInsert() {
-		encodePassword()
-	}
-
-	def beforeUpdate() {
-		if (isDirty('password')) {
-			encodePassword()
-		}
-	}
-
-	protected void encodePassword() {
-		password = springSecurityService?.passwordEncoder ? springSecurityService.encodePassword(password) : password
-	}
-
-	static transients = ['springSecurityService']
-
-	static constraints = {
-		username blank: false, unique: true, nullable: false
-		password blank: false, nullable: false
-		enabled nullable: false
-		accountExpired nullable: false
-		accountLocked nullable: false
-		passwordExpired nullable: false
-	}
-
-	static mapping = {
-		table 'TestUser'
-		password column: '`password`'
-	}
-}
