+++ grails-test-examples/spring-security/cas/test1/grails-app/domain/grails/plugin/springsecurity/cas/test/Role.groovy
@@ -17,8 +17,28 @@
  *  under the License.
  */
 
-import grails.plugin.springsecurity.cas.test.CasServiceUrlConfigurer
+package grails.plugin.springsecurity.cas.test
 
-beans = {
-    casServiceUrlConfigurer(CasServiceUrlConfigurer)
+import groovy.transform.EqualsAndHashCode
+import groovy.transform.ToString
+
+@EqualsAndHashCode(includes='authority')
+@ToString(includes='authority', includeNames=true, includePackage=false)
+class Role implements Serializable {
+
+	private static final long serialVersionUID = 1
+
+	String authority
+
+	Role(String authority) {
+		this.authority = authority
+	}
+
+	static constraints = {
+		authority blank: false, unique: true, nullable: false
+	}
+
+	static mapping = {
+		cache true
+	}
 }
